Remove utils.py in test folder and use scifem more
